Fix missing HA setup functions
