+++ __init__.py
@@ -1,0 +1,32 @@
+"""Home Assistant entry for the Special Agent integration."""
+
+from __future__ import annotations
+
+from homeassistant.core import HomeAssistant, ServiceCall
+from homeassistant.config_entries import ConfigEntry
+
+DOMAIN = "special_agent"
+PLATFORMS = ["conversation"]
+
+
+async def async_setup(hass: HomeAssistant, config: dict) -> bool:
+    """Set up via configuration.yaml (unused)."""
+
+    async def reload_service_handler(call: ServiceCall) -> None:
+        for entry in hass.config_entries.async_entries(DOMAIN):
+            await hass.config_entries.async_reload(entry.entry_id)
+
+    hass.services.async_register(DOMAIN, "reload", reload_service_handler)
+    return True
+
+
+async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
+    """Set up Special Agent from a config entry."""
+    hass.data.setdefault(DOMAIN, {})[entry.entry_id] = entry.data
+    await hass.config_entries.async_forward_entry_setups(entry, PLATFORMS)
+    return True
+
+
+async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
+    """Unload the integration."""
+    return await hass.config_entries.async_unload_platforms(entry, PLATFORMS)
